move tooltip regex to new file to prepare writing mappings
